main/requirement.txt
dnspython
psycopg2-binary
sqlalchemy
injector
starlette